# Improve performance of `ShaderProgram.fromCache()`