@ptrgags 2022-08-26
:construction: WIP :construction: 

One thing I wanted to test is that shaders are cached the same way between `1.96` and this branch. To do this, I'm trying a sampling of Sandcastles that print out the number of cache keys in the cache once it settles. Testing instructions:

1. Run the Sandcastle link and wait for the scene to load completely. Do not adjust the camera!
2. Hover over tilesets and click the canvas a few times to make sure any picking-related shaders have been added to the cache. 
3. After a few clicks the number should settle on a constant value, this is the value to compare with.

Number of shaders in cache:

| Sandcastle Link | `1.96` | This branch | 
|-----------------|--------|-------------|
| [CDB Yemen](http://localhost:8080/Apps/Sandcastle/index.html#c=pVZ9b9o4GP8qFjepqcSchCRAWTsdY2VD67RqsLs/jlNnnAfw5tjIduhY1e9+jxPStbTT7XQIlNjP+9vvIQzJUBH4xoqNBGLX+lqoFWFko4VyhEtd5sQJCRb8iVkrlgJystghzxvQBTizaxjoXHGtrCNbAddgyBlRcE1GYEVZ0D+qu2De4tV5pJVjQoGZt9rkZq4IcWAM3lwavRU5mEEjyA0wB39qI/NZzRIct+fq9vjFXM1VGF5Wno4qT9EtlF9ScgmojgylBNUmIy3LYiEY+aTEFowVbkc+6oV2gltywRaUTDlTygeOCoYSvjCVG01mRnOupSB4JO/RCUFGQnNmpIC7WJvkPAi2fiSvZzUxqAIsjbwLaqLVR7C6NBzo0uhiaJFtkgdxN+3EPjxCQLGFhPc6B3n+bQNGFIApQxXOlLCPv040tRwU0A2yCIcRWsryPNg71qTpR7UqpwgzgGX/obcK0q2BLLUpmCPCElsuvgDHEDXha6ZWQISqWUpXGp8CVDvWhhQaleWABZW2TSzAoCLhZ+3cxg7CcCXculxQrouwzsCbaZjkzysfQ2cAwi13JvSujSv7Nmz8bRLddB9nTmg1+09p909C5i1KQ/xOq2Z/zRwLH0iEowcm6s6qGitsOvyL1Qo71us78Ge3gbvqjh6R6Oh8Ovn0/ip5fTWbXJy3f6F+T8ZbV/NJErVuhyP804RMPblOCTa1xgG7qdOC6c2FV2QH5K/6iuBL69mNyG/J2dkZOTJaL+MjP6o4v144OPotitLxeHzUJhHNjuetv9tPizqNEb7SzuniUEOSvHrV7f6ahpnePHbg5GQ4/Fdx73vnf/ieHMiOx/1+kvyabPpIdjjsdJ6Wrcb6gHs8jqJH3PXz9gcGPugizgowjGJDeMSt652DdUJVzTJ40B7MOHxjKglq1Wkaxb00pd00jdMsidO9yU4ni7KYpnHU65/EcZR19oQ0O8nSTkSTXtrrp2mS+Osav7QRHlgeGX0LLEeovRSOrz9qKfemMzTa7SVJP8qi5KSX9fYWnke00+t2o16cJf3kpN/v7gld2unH6Eu/l6VRFPWzxvTdZiATVYMZukDYQpeuQi9eGoOeyR1Zi9Va4s/hRlvilkFQa9DmPumsnpQ9x4CUKocl7q68iXOFyZWjeqjup7eqY+VSPba1ZgRaebgdp7jlQE03jMP5Fp17WzMFB6VVW2Z9dHsdvsoTtSndkPsYg2WpqhcSFHoLHtePyX7gvWW7ZrhYR4yvPUw8UL002DZTrBZawc0M3xy9usf+4lCJRQX36A2zvWPUEqjUq+DzO9hZvze45xuQZzcfqqVCv+J9sJc6prinV259+9kHd9v+WVYqJL04H8+uRheT0bvjF61267SCvZfNIP0uio02zqN+gFDvAKEeo7LhouRfESS5rRLoWU/D+6KnudgSkZ898QelhnqkLEspp+I7junL0xD5H4lKXfX2B/yXIdnOs63jlxf1JaX0NMTj05JOa7lg5kDzPw) (Do not modify dropdowns)| 74 | TODO |
| same link, cycle through the 3 style options | 80 | TODO |
| [Point Cloud Classification](http://localhost:8080/Apps/Sandcastle/index.html#c=pVZ9b9o4GP8qFjepqcSchCRAWTsdY2VD67RqsLs/jlNnnAfw5tjIduhY1e9+jxPStbTT7XQIlNjP+9vvIQzJUBH4xoqNBGLX+lqoFWFko4VyhEtd5sQJCRb8iVkrlgJystghzxvQBTizaxjoXHGtrCNbAddgyBlRcE1GYEVZ0D+qu2De4tV5pJVjQoGZt9rkZq4IcWAM3lwavRU5mEEjyA0wB39qI/NZzRIct+fq9vjFXM1VGF5Wno4qT9EtlF9ScgmojgylBNUmIy3LYiEY+aTEFowVbkc+6oV2gltywRaUTDlTygeOCoYSvjCVG01mRnOupSB4JO/RCUFGQnNmpIC7WJvkPAi2fiSvZzUxqAIsjbwLaqLVR7C6NBzo0uhiaJFtkgdxN+3EPjxCQLGFhPc6B3n+bQNGFIApQxXOlLCPv040tRwU0A2yCIcRWsryPNg71qTpR7UqpwgzgGX/obcK0q2BLLUpmCPCElsuvgDHEDXha6ZWQISqWUpXGp8CVDvWhhQaleWABZW2TSzAoCLhZ+3cxg7CcCXculxQrouwzsCbaZjkzysfQ2cAwi13JvSujSv7Nmz8bRLddB9nTmg1+09p909C5i1KQ/xOq2Z/zRwLH0iEowcm6s6qGitsOvyL1Qo71us78Ge3gbvqjh6R6Oh8Ovn0/ip5fTWbXJy3f6F+T8ZbV/NJErVuhyP804RMPblOCTa1xgG7qdOC6c2FV2QH5K/6iuBL69mNyG/J2dkZOTJaL+MjP6o4v144OPotitLxeHzUJhHNjuetv9tPizqNEb7SzuniUEOSvHrV7f6ahpnePHbg5GQ4/Fdx73vnf/ieHMiOx/1+kvyabPpIdjjsdJ6Wrcb6gHs8jqJH3PXz9gcGPugizgowjGJDeMSt652DdUJVzTJ40B7MOHxjKglq1Wkaxb00pd00jdMsidO9yU4ni7KYpnHU65/EcZR19oQ0O8nSTkSTXtrrp2mS+Osav7QRHlgeGX0LLEeovRSOrz9qKfemMzTa7SVJP8qi5KSX9fYWnke00+t2o16cJf3kpN/v7gld2unH6Eu/l6VRFPWzxvTdZiATVYMZukDYQpeuQi9eGoOeyR1Zi9Va4s/hRlvilkFQa9DmPumsnpQ9x4CUKocl7q68iXOFyZWjeqjup7eqY+VSPba1ZgRaebgdp7jlQE03jMP5Fp17WzMFB6VVW2Z9dHsdvsoTtSndkPsYg2WpqhcSFHoLHtePyX7gvWW7ZrhYR4yvPUw8UL002DZTrBZawc0M3xy9usf+4lCJRQX36A2zvWPUEqjUq+DzO9hZvze45xuQZzcfqqVCv+J9sJc6prinV259+9kHd9v+WVYqJL04H8+uRheT0bvjF61267SCvZfNIP0uio02zqN+gFDvAKEeo7LhouRfESS5rRLoWU/D+6KnudgSkZ898QelhnqkLEspp+I7junL0xD5H4lKXfX2B/yXIdnOs63jlxf1JaX0NMTj05JOa7lg5kDzPw) | 59 | TODO |