@ptrgags 2022-08-26
To try a case that doesn't use `Model`/3D Tiles, I made [This billboard sandcastle](http://localhost:8080/Apps/Sandcastle/index.html#c=pVTBcpswEP2VHU5QOwKugXg6cY/t9JBOT1xkIfBOhUQl4Zh6/O8VEGHsOKdwQFrp7b630gOmpLFwQP7KNTyB5K+w5Qa7hvwe18IiYGO8VdJSlFwXQZQVspBVJ5lFJYGW5TMKsVNUl1slBB+XQ4zgNOAA2Mixe49xhBMzMYxLTlqNDVo8cENc0XDIhaWkezSnCQUwlni8Krie9s7RME6qvZpaY/mC/7iT8IPaPTF/tQ3TJBlQAJXSEApuoXf7SeaGfM7IYLXqx96G2jPyOCGPN8jjjPTMQskabVcO1G99jQoE123ooQAPabK+RFfBEWII5wYeII38XgQrQPgCCUmzG1ZqP0faf0x6Q9Uqg2+3u7i7LdVW1Zq2e2Sk0qr5xmvNuVmwzyezoPW6r2S5hyRLej+/47HRSadLtlf3eFeYVUPoNqgMPTJacGNDa2eyIiAkHucmrihDgbYnNVZFsMAaRoXDphepZ39U52Fwr/Mg/YXKklFjBR+0/lJK7Kh+7qx17i6CrebUcpitbxwHzN9e+Gav2YY42RCdDd05uclqNTvwww8183KiDIJ1kBvbC77xqr9i0yptodMidG1b3rTuVlznu4794ZYwY3xfebxMzUs8AJZPd34hwAQ1xu1UnRCDo4pgk8cO/y5VKFqirH8euBa0H2D7dPN9WiSE5LEL72fa6SRvKv8H) with a button that creates 100 `BillboardCollections`, each with 100 billboards every time the button is clicked.

My test procedure:

1. Load the sandcastle and rotate the globe to face `(lon, lat) = (0, 0)`. The exact view doesn't matter
2. Start the profiler
3. Press the Create Billboards button a total of 5 times
4. Stop the profiler after 10 seconds.

The performance difference wasn't quite as significant as `Model`, but I can confirm that `removeComments()` disappear from the top 10 list. Furthermore, the CPU usage seems to go back to idling levels faster on this branch.

| `main` | This branch |
| ------ | ------------|
|![2022-08-26_BillboardPerformanceMain](https://user-images.githubusercontent.com/8422414/186987495-687d8664-18c1-4262-b64f-6d35d05d48b8.png) | ![2022-08-26_BillboardPerformanceBranch](https://user-images.githubusercontent.com/8422414/186987524-295cf8d0-970a-42da-8357-568bb07687ea.png) |

The one thing I notice is a different shader-related function, `ShaderCache.getDerivedShaderProgram()` is still at the top (though it seems to be a smaller proportion of the runtime at least in these runs). This function only involves string concatenation and a dictionary lookup, it doesn't do any expensive shader compilation like `ShaderProgram.fromCache()`. Though bear in mind, this is a contrived use case that's creating far too many `BillboardCollections`, several hundred cache lookups for only ~36 shaders in the cache.