@ptrgags 2022-08-26
@j9liu this is ready for review! With the above I tried a variety of different cases and this change is working nicely across the board